+++ public/js/helper.js
@@ -177,3 +177,15 @@ function appendLink(href, rel = "stylesheet") {
     document.head.appendChild(link);
   }
 }
+
+function openCalendly() {
+  appendScript("https://assets.calendly.com/assets/external/widget.js", true);
+  appendLink("https://assets.calendly.com/assets/external/widget.css");
+  setTimeout(() => {
+    if (window.Calendly) {
+      window.Calendly.initPopupWidget({
+        url: 'https://calendly.com/sales-accounting-software/talk-to-sale',
+      }, 1000);
+    };
+  })
+}
