+++ src/components/bannerSection.js
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { createSlider } from "@typeform/embed";
 import Head from 'next/head';
 import Jsondata from '../data/schema/organizationSchema.json';
@@ -9,14 +9,6 @@ const bannerSection = (path) => {
   const [utm, setUtm] = useState(null);
   const selectedSchema = getSchemaForLinkPath(link, Jsondata.organizationSchema);
 
-//  useEffect(() => {
-//    const utmParams = new URLSearchParams({
-//      utm_source: getLocalStorage("utm_source"),
-//      utm_medium: getLocalStorage("utm_medium"),
-//      utm_campaign: getLocalStorage("utm_campaign"),
-//    }).toString();
-//    setUtm(utmParams);
-//  }, [utm]);
 
     const openTypeForm = () => {
       const utmParams = new URLSearchParams({
