+++ src/pages/cloud-accounting-software.js
@@ -15,18 +15,6 @@ const anywhereAnytime = (path) => {
   const isUK = linkPath.isUK;
   const link = linkPath.linkPrefix;
 
-  function openCalendly(){
-    appendScript("https://assets.calendly.com/assets/external/widget.js", true);
-    appendLink("https://assets.calendly.com/assets/external/widget.css");
-    setTimeout(() => {
-      if (window.Calendly) {
-        window.Calendly.initPopupWidget({
-          url: 'https://calendly.com/sales-accounting-software/talk-to-sale',
-        }, 500);
-      };
-    })
-  }
-
   useEffect(() => {
     setFaq(faqs['cloud-accounting-software'])
   }, []);
@@ -75,7 +63,7 @@ const anywhereAnytime = (path) => {
                     src="/img/feature-gallery-icons/cloud-banner.svg"
                     alt="Illustration of cloud accounting software"
                     width="400"
-                    height="300"
+                    height="auto"
                   />
                   <img
                     src="/img/all-features-cloud_icon.svg"
@@ -157,7 +145,7 @@ const anywhereAnytime = (path) => {
                     className="img-fluid"
                     alt="connect bank image"
                     width="400"
-                    height="400"
+                    height="auto"
                   />
                 </figure>
               </div>
@@ -206,7 +194,7 @@ const anywhereAnytime = (path) => {
                     className="img-fluid"
                     alt="bank reconcile image"
                     width="400"
-                    height="400"
+                    height="auto"
                   />
                 </figure>
               </div>
@@ -255,7 +243,7 @@ const anywhereAnytime = (path) => {
                     className="img-fluid"
                     alt="bank account image"
                     width="400"
-                    height="400"
+                    height="auto"
                   />
                 </figure>
               </div>
@@ -289,7 +277,7 @@ const anywhereAnytime = (path) => {
                     className="img-fluid"
                     alt="bank overview image"
                     width="400"
-                    height="400"
+                    height="auto"
                   />
                 </figure>
               </div>
@@ -342,7 +330,7 @@ const anywhereAnytime = (path) => {
                         className="img-fluid"
                         alt="bank account image"
                         width="400"
-                        height="400"
+                        height="auto"
                       />
                     </figure>
                   </div>
@@ -479,7 +467,7 @@ const anywhereAnytime = (path) => {
                         className="img-fluid"
                         alt="bank overview image"
                         width="400"
-                        height="400"
+                        height="auto"
                       />
                     </figure>
                   </div>
@@ -742,7 +730,7 @@ const anywhereAnytime = (path) => {
                     alt="Illustration representing suggestions for Cloud Accounting Software"
                     className="features__sub_heading_container--banner_img"
                     width="200"
-                    height="200"
+                    height="auto"
                   />
                 </figure>
                 <h2 id="suggestion-heading" className="c-fs-3 col-grey-deep c-fw-400 mb-3 mt-4">
