+++ src/pages/conect-bank-reconcile.js
@@ -8,18 +8,6 @@ const conectBankReconcile = (path) => {
   const isIndia = linkPath.isIndia
   const link = linkPath.linkPrefix;
   const [faq, setFaq] = useState([])
-  
-  function openCalendly() {
-    appendScript("https://assets.calendly.com/assets/external/widget.js", true);
-    appendLink("https://assets.calendly.com/assets/external/widget.css");
-    setTimeout(()=>{
-      if (window.Calendly) {
-        window.Calendly.initPopupWidget({
-          url: 'https://calendly.com/sales-accounting-software/talk-to-sale',
-        }, 500)
-      }
-    })
-  }
 
   useEffect(() => {
     setFaq(faqs['conect-bank-reconcile']);
@@ -133,7 +121,7 @@ const conectBankReconcile = (path) => {
                       className="img-fluid"
                       alt="Illustration showing the process of connecting a bank account"
                       width="300"
-                      height="300"
+                      height="auto"
                     />
                   </figure>
                 </div>
@@ -161,7 +149,7 @@ const conectBankReconcile = (path) => {
                     className="img-fluid"
                     alt="Illustration depicting the bank reconciliation process"
                     width="300"
-                    height="300"
+                    height="auto"
                   />
                 </figure>
               </div>
@@ -188,7 +176,7 @@ const conectBankReconcile = (path) => {
                     className="img-fluid"
                     alt="Illustration showing multiple bank accounts"
                     width="300"
-                    height="300"
+                    height="auto"
                   />
                 </figure>
               </div>
@@ -206,7 +194,7 @@ const conectBankReconcile = (path) => {
                     alt="Illustration of people interested in Connect Bank Reconcile"
                     className="features__sub_heading_container--banner_img"
                     width="200"
-                    height="200"
+                    height="auto"
                   />
                 </figure>
                 <h2 id="suggestion-heading" className="c-fs-3 col-grey-deep c-fw-400 mb-3 mt-4">
