+++ src/pages/contact-us.js
@@ -11,27 +11,6 @@ const contactUs = (path) => {
   const isUK = linkPath.isUK;
   const selectedSchema = getSchemaForLinkPath(linkPath, Jsondata.organizationSchema);
 
-    // useEffect(() => {
-    //   const utmParams = new URLSearchParams({
-    //     utm_source: getLocalStorage("utm_source"),
-    //     utm_medium: getLocalStorage("utm_medium"),
-    //     utm_campaign: getLocalStorage("utm_campaign"),
-    //   }).toString();
-    //   setUtm(utmParams);
-    // }, [utm]);
-
-    function openCalendly() {
-      appendScript("https://assets.calendly.com/assets/external/widget.js", true);
-      appendLink("https://assets.calendly.com/assets/external/widget.css");
-      setTimeout(() => {
-        if (window.Calendly) {
-          window.Calendly.initPopupWidget({
-            url: 'https://calendly.com/sales-accounting-software/talk-to-sale',
-          }, 500);
-        };
-      })
-    }
-
    const openTypeForm = () => {
     const utmParams = new URLSearchParams({
       utm_source: getLocalStorage("utm_source"),
