+++ src/pages/contacts.js
@@ -4,18 +4,6 @@ const contacts = (path) => {
   const isIndia = linkPath.isIndia
   const isAE = linkPath.isAE;
   const link = linkPath.linkPrefix;
-
-  function openCalendly() {
-    appendScript("https://assets.calendly.com/assets/external/widget.js", true);
-    appendLink("https://assets.calendly.com/assets/external/widget.css");
-    setTimeout(() => {
-      if (window.Calendly) {
-        window.Calendly.initPopupWidget({
-          url: 'https://calendly.com/sales-accounting-software/talk-to-sale',
-        }, 500);
-      };
-    })
-  }
   
   return (
     <>
@@ -121,8 +109,8 @@ const contacts = (path) => {
                     src="/img/feature-gallery-icons/import-data-feature-1.webp"
                     className="img-fluid"
                     alt="Illustration of connecting bank data"
-                    width="400"
-                    height="400"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -147,8 +135,8 @@ const contacts = (path) => {
                     src={ isIndia ? "/img/feature-gallery-icons/reporting-feature-3.webp" : ( isAE ? "/img/feature-gallery-icons/uk-aging-report.svg" : "/img/feature-gallery-icons/uk-aging-report.svg") }
                     className="img-fluid"
                     alt="Aging report illustration based on region"
-                    width="400"
-                    height="400"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -184,8 +172,8 @@ const contacts = (path) => {
                     src="/img/feature-gallery-icons/import-data-feature-4.svg"
                     className="img-fluid"
                     alt="Illustration of bank account management"
-                    width="400"
-                    height="400"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -202,7 +190,7 @@ const contacts = (path) => {
                     src="/img/suggestion.svg"
                     alt="Illustration of people interested in Connect Bank Reconcile"
                     className="features__sub_heading_container--banner_img"
-                    width="200"
+                    width="auto"
                     height="200"
                   />
                 </figure>
