+++ src/pages/in/e-invoice.js
@@ -8,18 +8,6 @@ const eInvoice = () => {
   useEffect(() => {
     setFaq(faqs['e-invoice']);
   }, []);
-  
-  function openCalendly() {
-    appendScript("https://assets.calendly.com/assets/external/widget.js", true);
-    appendLink("https://assets.calendly.com/assets/external/widget.css");
-    setTimeout(() => {
-      if (window.Calendly) {
-          window.Calendly.initPopupWidget({
-              url: 'https://calendly.com/sales-accounting-software/talk-to-sale',
-          }, 500);
-      };
-    })
-  }
 
   return (
     <div className="e-invoice">
