+++ src/pages/inventory-management-software.js
@@ -11,18 +11,6 @@ const inventoryManagementSoftware = (path) => {
   const link = linkPath.linkPrefix;
   const [faq, setFaq]= useState([])
 
-  function openCalendly() {
-    appendScript("https://assets.calendly.com/assets/external/widget.js", true);
-    appendLink("https://assets.calendly.com/assets/external/widget.css");
-    setTimeout(() => {
-      if (window.Calendly) {
-        window.Calendly.initPopupWidget({
-          url: 'https://calendly.com/sales-accounting-software/talk-to-sale',
-        }, 500);
-      };
-    })
-  }
-
   useEffect(() => {
     setFaq(faqs['inventory-management-software']);
   }, []);
@@ -127,8 +115,8 @@ const inventoryManagementSoftware = (path) => {
                     src="/img/feature-gallery-icons/record-inventory.svg"
                     className="img-fluid"
                     alt="Illustration of recording inventory for products and services"
-                    width="300"
-                    height="300"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -152,8 +140,8 @@ const inventoryManagementSoftware = (path) => {
                     src="/img/feature-gallery-icons/inventory-master.svg"
                     className="img-fluid"
                     alt="Illustration of inventory masters"
-                    width="300"
-                    height="300"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -178,8 +166,8 @@ const inventoryManagementSoftware = (path) => {
                     src="/img/feature-gallery-icons/bifurcate-inventory.svg"
                     className="img-fluid"
                     alt="Illustration of bifurcating inventory"
-                    width="300"
-                    height="300"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -206,8 +194,8 @@ const inventoryManagementSoftware = (path) => {
                     src="/img/feature-gallery-icons/reports.svg"
                     className="img-fluid"
                     alt="Illustration of real-time inventory reports"
-                    width="300"
-                    height="300"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -235,8 +223,8 @@ const inventoryManagementSoftware = (path) => {
                     src="/img/feature-gallery-icons/barcode-scanning.svg"
                     className="img-fluid"
                     alt="barcode-scanning-feature-icon"
-                    width="300"
-                    height="300"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -263,8 +251,8 @@ const inventoryManagementSoftware = (path) => {
                     src="/img/feature-gallery-icons/manufacturing-export.svg"
                     className="img-fluid"
                     alt="bank reconcile image"
-                    width="300"
-                    height="300"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -292,8 +280,8 @@ const inventoryManagementSoftware = (path) => {
                     src="/img/feature-gallery-icons/branch-warehouse.svg"
                     className="img-fluid"
                     alt="bank reconcile image"
-                    width="300"
-                    height="300"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -319,8 +307,8 @@ const inventoryManagementSoftware = (path) => {
                     src="/img/feature-gallery-icons/branch-transfer.svg"
                     className="img-fluid"
                     alt="bank account image"
-                    width="300"
-                    height="300"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -345,8 +333,8 @@ const inventoryManagementSoftware = (path) => {
                     src="/img/feature-gallery-icons/custom-unit.svg"
                     className="img-fluid"
                     alt="custom-units-icon"
-                    width="300"
-                    height="300"
+                    width="auto"
+                    height="200"
                   />
                 </figure>
               </div>
@@ -363,7 +351,7 @@ const inventoryManagementSoftware = (path) => {
                     src="/img/suggestion.svg"
                     alt="Illustration of people interested in inventory management software"
                     className="features__sub_heading_container--banner_img"
-                    width="200"
+                    width="auto"
                     height="200"
                   />
                 </figure>
